+++ src/index.html
@@ -9,7 +9,7 @@
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css">
     <title>Namaste React Bootcamp</title>
 </head>
-<body class="app-body">
+<body>
     <div id="react-root">Not Rendered</div>
     <script type="module" src="./App.js"></script>
 </body>
