+++ src/style.css
@@ -1,5 +1,5 @@
-.app-body{
-    background-color: #d0b6da;
+body{
+    background-color: hsl(283, 48%, 83%);
 }
 
 /** styles for header component */
